requirements.txt
aiofiles==23.2.1
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.0.4
clr-loader==0.2.6
configobj==5.0.8
configparser==6.0.0
docbarcodes==1.0.9
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
future==0.18.3
h11==0.14.0
httplib2==0.22.0
hypothesis==6.96.2
idna==3.6
iniconfig==2.0.0
IronPdf==2023.8.6
isodate==0.6.1
itsdangerous==2.1.2
JPype1==1.3.0
loguru==0.5.3
looseversion==1.3.0
lxml==5.1.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.3
opencv-python==4.5.4.60
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pdf2image==1.16.0
pdf417decoder==1.0.8
pillow==10.2.0
pluggy==1.3.0
prov==2.0.0
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pydot==2.0.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytest==6.2.5
pytest-datadir==1.5.0
pytest-regressions==2.2.0
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
pyzbar==0.1.9
rdflib==7.0.0
requests==2.31.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.35.1
toml==0.10.2
traits==6.3.2
typer==0.4.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
zxingmod==0.19
